Add finnhub-secondary EA (#2855)
* Finnhub: Add WS Transport

Adds WebSocket Transport to Quote endpoint. Default transport remains REST.

* Finnhub: Expose builder for quote endpoint

Exposes a function for building the quote endpoint based on provided overrides. This will be used by the new finnhub-secondary EA, which will run the same EA with different overrides.

* Add finnhub-secondary EA

Add new finnhub-secondary EA, which is the same as the finnhub EA but with different default overrides to support OANDA